requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
authlib==0.15.2           # via -r requirements.in
certifi==2020.11.8        # via elasticsearch
cffi==1.14.3              # via cryptography
click==7.1.2              # via flask
cryptography==3.2.1       # via authlib
elasticsearch==7.10.0     # via -r requirements.in
flask-wtf==0.14.3         # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
itsdangerous==1.1.0       # via flask, flask-wtf
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2, wtforms
pycparser==2.20           # via cffi
python-dotenv==0.15.0     # via -r requirements.in
six==1.15.0               # via cryptography
urllib3==1.26.2           # via elasticsearch
werkzeug==1.0.1           # via flask
wtforms==2.3.3            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
